# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /model/v4/projections:
    get:
      operationId: getV4FullSlateProjections
      tags: [model]
      summary: Get safe full-slate V4 model projections
      description: >-
        Returns every legitimate V4 forecast for one sport/day plus explicit
        coverage failures. Validating projections are not official TBM picks.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sport
          in: query
          required: true
          schema:
            type: string
            enum: [MLB, NCAAF, NFL, NBA, WNBA, NHL, SOCCER, UFC, NCAAMB]
        - name: date
          in: query
          required: false
          description: YYYY-MM-DD in America/New_York; omitted selects today.
          schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
      responses:
        "200":
          description: Safe model projections and full-slate coverage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/V4FullSlateProjectionResponse" }
        "400": { description: Invalid sport or date }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    masterApiKey:
      type: apiKey
      in: header
      name: X-Master-Key
      description: Internal master key. An authenticated admin session token is also accepted by the server.
  schemas:
    V4FullSlateProjectionResponse:
      type: object
      required: [sport, date, lifecycleDisclaimer, coverage, officialPicks, projections]
      properties:
        sport: { type: string }
        date: { type: string }
        lifecycleDisclaimer: { type: string }
        coverage:
          type: object
          required: [scheduledEvents, eligibleEvents, forecastedEvents, failedEvents, forecastCoveragePct, failures]
          properties:
            scheduledEvents: { type: integer }
            eligibleEvents: { type: integer }
            forecastedEvents: { type: integer }
            failedEvents: { type: integer }
            forecastCoveragePct: { type: number }
            failures:
              type: array
              items:
                type: object
                required: [gameId, reason]
                properties:
                  gameId: { type: string }
                  reason: { type: string }
        officialPicks:
          type: array
          description: Persisted server-authoritative official V4 decisions; independent of current projection IDs.
          items: { $ref: "#/components/schemas/V4OfficialPick" }
        projections:
          type: array
          items: { $ref: "#/components/schemas/V4PublicProjection" }
    V4PublicProjection:
      type: object
      required: [eventId, sport, modelVersion, lifecycleStatus, forecastTimestamp, officialPickStatus, officialRole, officialRank, units, officialFailureReason]
      properties:
        eventId: { type: string }
        sport: { type: string }
        eventStart: { type: [string, "null"] }
        homeParticipant: { type: [string, "null"] }
        awayParticipant: { type: [string, "null"] }
        modelVersion: { type: string }
        lifecycleStatus:
          type: string
          enum: [V4_VALIDATING, V4_PROVISIONAL, V4_APPROVED]
        expectedHomeScore: { type: [number, "null"] }
        expectedAwayScore: { type: [number, "null"] }
        expectedMargin: { type: [number, "null"] }
        expectedTotal: { type: [number, "null"] }
        homeWinProbability: { type: [number, "null"] }
        drawProbability: { type: [number, "null"] }
        awayWinProbability: { type: [number, "null"] }
        projectedWinner: { type: [string, "null"], enum: [HOME, DRAW, AWAY, null] }
        forecastTimestamp: { type: string }
        officialPickStatus:
          type: string
          enum: [NOT_PUBLICATION_ELIGIBLE, NO_OFFICIAL_PLAY, OFFICIAL_TBM_PLAY]
        officialRole:
          type: string
          description: Server-authoritative public role. Clients must not recompute rank or POTD.
          enum: [TOP_PLAY, QUALIFIED_PLAY, PROJECTION]
        officialRank:
          type: [integer, "null"]
          minimum: 1
          description: Server-authoritative rank among qualified V4 plays; null for projections.
        units:
          type: integer
          enum: [0, 1]
          description: Official qualified plays are always flat 1U.
        officialFailureReason:
          type: [string, "null"]
          description: Explicit fail-closed reason when no official V4 play exists.
    V4OfficialPick:
      type: object
      required: [eventId, sport, market, selection, role, rank, units, status]
      properties:
        eventId: { type: string }
        sport: { type: string }
        homeParticipant: { type: [string, "null"] }
        awayParticipant: { type: [string, "null"] }
        market: { type: string }
        selection: { type: string }
        odds: { type: [integer, "null"] }
        modelProbability: { type: [number, "null"] }
        fairProbability: { type: [number, "null"] }
        role: { type: string, enum: [TOP_PLAY, QUALIFIED_PLAY] }
        rank: { type: integer, minimum: 1 }
        units: { type: integer, enum: [1] }
        status: { type: string, enum: [PUBLISHED] }
        modelId: { type: [string, "null"] }
        modelVersion: { type: [string, "null"] }
        artifactId: { type: [string, "null"] }
        artifactHash: { type: [string, "null"] }
        inputHash: { type: [string, "null"] }
        marketEvidenceId: { type: [string, "null"] }
...