# lib/api-zod/src/generated/types/v4FullSlateProjectionResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { V4FullSlateProjectionResponseCoverage } from './v4FullSlateProjectionResponseCoverage';
import type { V4OfficialPick } from './v4OfficialPick';
import type { V4PublicProjection } from './v4PublicProjection';

export interface V4FullSlateProjectionResponse {
  sport: string;
  date: string;
  lifecycleDisclaimer: string;
  coverage: V4FullSlateProjectionResponseCoverage;
  /** Persisted server-authoritative official V4 decisions; independent of current projection IDs. */
  officialPicks: V4OfficialPick[];
  projections: V4PublicProjection[];
}