# lib/api-zod/src/generated/types/v4OfficialPick.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { V4OfficialPickRole } from './v4OfficialPickRole';
import type { V4OfficialPickStatus } from './v4OfficialPickStatus';
import type { V4OfficialPickUnits } from './v4OfficialPickUnits';

export interface V4OfficialPick {
  eventId: string;
  sport: string;
  /** @nullable */
  homeParticipant?: string | null;
  /** @nullable */
  awayParticipant?: string | null;
  market: string;
  selection: string;
  /** @nullable */
  odds?: number | null;
  /** @nullable */
  modelProbability?: number | null;
  /** @nullable */
  fairProbability?: number | null;
  role: V4OfficialPickRole;
  /** @minimum 1 */
  rank: number;
  units: V4OfficialPickUnits;
  status: V4OfficialPickStatus;
  /** @nullable */
  modelId?: string | null;
  /** @nullable */
  modelVersion?: string | null;
  /** @nullable */
  artifactId?: string | null;
  /** @nullable */
  artifactHash?: string | null;
  /** @nullable */
  inputHash?: string | null;
  /** @nullable */
  marketEvidenceId?: string | null;
}