# lib/api-zod/src/generated/types/v4PublicProjection.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { V4PublicProjectionLifecycleStatus } from './v4PublicProjectionLifecycleStatus';
import type { V4PublicProjectionOfficialPickStatus } from './v4PublicProjectionOfficialPickStatus';
import type { V4PublicProjectionOfficialRole } from './v4PublicProjectionOfficialRole';
import type { V4PublicProjectionProjectedWinner } from './v4PublicProjectionProjectedWinner';
import type { V4PublicProjectionUnits } from './v4PublicProjectionUnits';

export interface V4PublicProjection {
  eventId: string;
  sport: string;
  /** @nullable */
  eventStart?: string | null;
  /** @nullable */
  homeParticipant?: string | null;
  /** @nullable */
  awayParticipant?: string | null;
  modelVersion: string;
  lifecycleStatus: V4PublicProjectionLifecycleStatus;
  /** @nullable */
  expectedHomeScore?: number | null;
  /** @nullable */
  expectedAwayScore?: number | null;
  /** @nullable */
  expectedMargin?: number | null;
  /** @nullable */
  expectedTotal?: number | null;
  /** @nullable */
  homeWinProbability?: number | null;
  /** @nullable */
  drawProbability?: number | null;
  /** @nullable */
  awayWinProbability?: number | null;
  /** @nullable */
  projectedWinner?: V4PublicProjectionProjectedWinner;
  forecastTimestamp: string;
  officialPickStatus: V4PublicProjectionOfficialPickStatus;
  /** Server-authoritative public role. Clients must not recompute rank or POTD. */
  officialRole: V4PublicProjectionOfficialRole;
  /**
     * Server-authoritative rank among qualified V4 plays; null for projections.
     * @minimum 1
     * @nullable
     */
  officialRank: number | null;
  /** Official qualified plays are always flat 1U. */
  units: V4PublicProjectionUnits;
  /**
     * Explicit fail-closed reason when no official V4 play exists.
     * @nullable
     */
  officialFailureReason: string | null;
}